Preserve header directory structure on install
CMake: Use a file set on the target, not only properly makes CMake
aware of the headers, it also sets the include directory properly
and preserves the subfolder when installing the file set.

Meson: Install headers with preserve_path to preserve subfolders.

The includes in the header files were changed also to properly
reference the other headers inside the correct subdirectories.