CMake: Read API version from version header
This prevents them getting out of sync and allows us to have a
single source of truth.

CMake does not have an easy way to get a preprocessor define from a
file so we have to use a regex, but as we don't have recursive expansion
here, this is fine.